Implement video upload flow with presigned URLs and processing status
