+++ apps/mobile/package.json
@@ -13,6 +13,13 @@
     "expo": "~51.0.17",
     "expo-av": "~14.0.6",
     "expo-haptics": "~12.0.1",
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/bottom-tabs": "^6.5.20",
+    "react-native-safe-area-context": "^4.10.5",
+    "react-native-screens": "^3.31.1",
+    "react-native-gesture-handler": "^2.20.2",
+    "react-native-reanimated": "^3.10.1",
+    "expo-document-picker": "~12.0.2",
     "react": "18.2.0",
     "react-native": "0.74.3"
   },
